Creates and added an ISC example
Uses the "new" style: no console or additional functions in the main example, just find device, connect, and move. Need to adjust this example as it accounts for 4+ devices
