+++ C#/KCube/KIM101/Program.cs
@@ -11,17 +11,12 @@ namespace KIM_Console_net_managed
     {
         static void Main(string[] args)
         {
-            // Get parameters from command line
-            int argc = args.Count();
-            if (argc < 1)
-            {
-                Console.WriteLine("Usage: KIM_Console_net_managed serial_number");
-                Console.ReadKey();
-                return;
-            }
+            // Uncomment this line (and SimulationManager.Instance.UninitializeSimulations() at the end on Main)
+            // If you are using a simulated device
+            // SimulationManager.Instance.InitializeSimulations();
 
-            // Get the serial number (e.g. 97000123)
-            string serialNo = args[0];
+            // Enter the serial number for your device
+            string serialNo = "97000001";
 
             try
             {
@@ -121,6 +116,9 @@ namespace KIM_Console_net_managed
             // Tidy up and exit
             device.StopPolling();
             device.Disconnect(true);
+
+            // Uncomment this line if you are using Simulations
+            //SimulationManager.Instance.UninitializeSimulations();
 
             Console.ReadKey();
         }
