+++ MotionControl_Examples.sln
@@ -21,7 +21,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "CSharp", "CSharp", "{B96446
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "KCube", "KCube", "{0494B9C3-AE9B-42EE-875A-603FD2197C3E}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "KDB101", "C++\KCube\KDB101\KDB101.vcxproj", "{5AC44299-B705-44C9-906F-72ECCE54E2D8}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "KBD101", "C++\KCube\KDB101\KDB101.vcxproj", "{5AC44299-B705-44C9-906F-72ECCE54E2D8}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "KLD101", "C++\KCube\KLD101\KLD101.vcxproj", "{23C9E506-4E9F-49EB-8B27-B71F046B9EA0}"
 EndProject
@@ -141,6 +141,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		LICENSE = LICENSE
 		README.md = README.md
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ISC", "C#\Benchtop\ISC\ISC\ISC.csproj", "{B65B845D-521B-40A7-8630-3FEA69E8D53B}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -752,6 +754,18 @@ Global
 		{9B624893-CBE6-4F74-96AA-1135C6864490}.Release|x64.Build.0 = Release|Any CPU
 		{9B624893-CBE6-4F74-96AA-1135C6864490}.Release|x86.ActiveCfg = Release|Any CPU
 		{9B624893-CBE6-4F74-96AA-1135C6864490}.Release|x86.Build.0 = Release|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Debug|x64.Build.0 = Debug|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Debug|x86.Build.0 = Debug|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Release|x64.ActiveCfg = Release|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Release|x64.Build.0 = Release|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Release|x86.ActiveCfg = Release|Any CPU
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -818,6 +832,7 @@ Global
 		{A77C8DCD-1185-4C1D-BA53-CB70031D6FAE} = {070FCEC2-3619-474E-9E1C-B921287F7F6F}
 		{2C73C145-9BAD-4A1A-87A2-353498F98276} = {070FCEC2-3619-474E-9E1C-B921287F7F6F}
 		{9B624893-CBE6-4F74-96AA-1135C6864490} = {070FCEC2-3619-474E-9E1C-B921287F7F6F}
+		{B65B845D-521B-40A7-8630-3FEA69E8D53B} = {434EC6FB-2905-43A6-B80C-24627BDBAC9E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {0E819AF5-C7CF-4B48-9F7A-75C053699079}
